// src/app/shared/interfaces/arquivos.model.ts
export interface ArquivoMV {
  descricao?: string;
  extensao?: string;
  nomeDocumento?: string;
  tipo?: string;
  entidade?: string;
  url?: string;
  href?: string;
  safeURL?: any;
  pdfURL?: any;
  isDoc?: boolean;
  isImage?: boolean;
  isVideo?: boolean;
  descricaoFormatada?: string;
  dataAdicao?: string;
  id?: number;
  base64?: string;
  size?: number;
}